Breaking ReportController into its own file
