+++ INZFS.MVC/Views/Admin/GetApplicationsSearch.cshtml
@@ -49,4 +49,4 @@
     </tbody>
 
 </table>
-<button asp-controller="Admin" asp-action="GeneratePdf">Download Application</button>
+<button asp-controller="Report" asp-action="GeneratePdf">Download Application</button>
